+++ src/providers/comscore/comscore.js
@@ -1,8 +1,8 @@
-// Comscore
+// comScore
 // ---------
 // [Documentation](http://direct.comscore.com/clients/help/FAQ.aspx#faqTagging)
 
-analytics.addProvider('Comscore', {
+analytics.addProvider('comScore', {
 
     settings : {
         c1 : '2',